Fix use of uninitialized is_subgrid_var variable in init_output_fields routine
In the init_output_fields routine, the is_subgrid_var local variable was
previously being used before it was initialized (and in fact, that variable was
never set anywhere in the routine).

The ultimate effect of the use of the uninitialized is_subgrid_var variable was
a lack of output variables being defined in metgrid output files, though this
effect was observable only under certain conditions (e.g., with GNU compiler
versions >= 14.2.0, for example).